Use an OpenApi spec for the Portal API
The addition of the spec should make it easier for people building their
own systems against the Portal API. The spec is not complete, but
contains enough to be useful.

I note that this won't run on Ruby later than 2.3.
The issue is the vCloud-rest gem brings loads an old version of
rest-client (from 2015), that is incompatible with modern versions of
Ruby.